# connectors/docs/swagger.yaml
definitions:
...
  irminmodels.IrminAPIPaginationMetadata:
    properties:
      has_more:
        description: HasMore indicates if there are more items available
        example: true
        type: boolean
      next:
        description: Next is the next page number or token, if applicable
        example: "3"
        type: string
      page:
        description: Current page number
        example: 2
        minimum: 1
        type: integer
      per_page:
        description: Number of items per page
        example: 20
        minimum: 1
        type: integer
      total:
        description: Total is the total number of items available
        example: 150
        minimum: 0
        type: integer
      total_pages:
        description: Total number of pages available
        example: 8
        minimum: 0
        type: integer
    type: object
  irminmodels.IrminAPIResponse:
    properties:
      data: {}
      errors:
        example:
        - Field validation failed
        - Invalid input format
        items:
          type: string
        type: array
      message:
        example: Operation completed successfully
        maxLength: 140
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      pagination:
        $ref: '#/definitions/irminmodels.IrminAPIPaginationMetadata'
    type: object
...
host: connectors.irmin.co
info:
  contact:
    email: hello@irmin.co
    name: Irmin Team
    url: https://irmin.co
  description: Collection of connectors for Irmin to interact with external services
  termsOfService: http://irmin.co/terms
  title: Irmin Connectors
  version: 1.0.0
paths:
...
  /swagger:
    get:
      consumes:
      - application/json
      description: Interactive API documentation interface using Swagger UI
      produces:
      - text/html
      responses:
        "200":
          description: Swagger UI HTML page
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/irminmodels.IrminAPIResponse'
      summary: Swagger UI interface
      tags:
      - documentation
  /swagger.json:
    get:
      consumes:
      - application/json
      description: Retrieve the OpenAPI 3.0 specification in JSON format for this
        API
      produces:
      - application/json
      responses:
        "200":
          description: OpenAPI 3.0 specification in JSON format
          schema:
            type: object
        "404":
          description: Swagger specification file not found
          schema:
            $ref: '#/definitions/irminmodels.IrminAPIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/irminmodels.IrminAPIResponse'
      summary: Get Swagger JSON specification
      tags:
      - documentation
schemes:
- https
securityDefinitions:
  OperationTokenAuth:
    description: Operation token with Bearer prefix - received from operation/init
      and used for data operation endpoints
    in: header
    name: Authorization
    type: apiKey
  SystemTokenAuth:
    description: System token with Bearer prefix - used for connector registration
      and initialization endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"